Separating MethodAUpdater and PrefixUpdater into their own files
