+++ commons/src/main/java/fr/free/nrw/commons/category/AsyncResponse.java
@@ -1,0 +1,4 @@
+package fr.free.nrw.commons.category;
+
+public interface AsyncResponse {
+}
